% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.rrr}
\alias{plot.sofar}
\alias{plot.cv.sofar}
\alias{plot.srrr}
\alias{plot.cv.srrr}
\alias{plot.rssvd}
\title{Scatter Plot}
\usage{
\method{plot}{rrr}(x, y = NULL, layer = 1L,
  xlab = paste("latent predictor ", layer, sep = ""),
  ylab = paste("latent response ", layer, sep = ""), ...)

\method{plot}{sofar}(x, y = NULL, layer = 1L,
  xlab = paste("latent predictor ", layer, sep = ""),
  ylab = paste("latent response ", layer, sep = ""), ...)

\method{plot}{cv.sofar}(x, y = NULL, layer = 1L,
  xlab = paste("latent predictor ", layer, sep = ""),
  ylab = paste("latent response ", layer, sep = ""), ...)

\method{plot}{srrr}(x, y = NULL, layer = 1L,
  xlab = paste("latent predictor ", layer, sep = ""),
  ylab = paste("latent response ", layer, sep = ""), ...)

\method{plot}{cv.srrr}(x, y = NULL, layer = 1L,
  xlab = paste("latent predictor ", layer, sep = ""),
  ylab = paste("latent response ", layer, sep = ""), ...)

\method{plot}{rssvd}(x, y = NULL, layer = 1L,
  xlab = paste("latent predictor ", layer, sep = ""),
  ylab = paste("latent response ", layer, sep = ""), ...)
}
\arguments{
\item{x}{Some object generated by \code{rrpack}.}

\item{y}{NULL. Do not need to specify.}

\item{layer}{The unit-rank layer to plot; cannot be larger than the estimated rank}

\item{xlab}{Label of X axis.}

\item{ylab}{Label of Y axis.}

\item{...}{Other argumnts for future usage.}
}
\value{
ggplot2 object.
}
\description{
S3 methods generating scatter plot for some objects generated by
\code{rrpack} using \code{ggplot2}. An \code{ggplot2} object is returned so
that users are allowed to easily further customize the plot.
}
