% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{rrpack-coef}
\alias{rrpack-coef}
\alias{coef.mrrr}
\alias{coef.cv.mrrr}
\alias{coef.r4}
\alias{coef.rrr}
\alias{coef.rrr.fit}
\alias{coef.cv.rrr}
\alias{coef.srrr}
\alias{coef.sofar}
\alias{coef.rssvd}
\title{Estimated coefficients}
\usage{
\method{coef}{mrrr}(object, ...)

\method{coef}{cv.mrrr}(object, ...)

\method{coef}{r4}(object, ...)

\method{coef}{rrr}(object, ...)

\method{coef}{rrr.fit}(object, ...)

\method{coef}{cv.rrr}(object, ...)

\method{coef}{srrr}(object, ...)

\method{coef}{sofar}(object, ...)

\method{coef}{rssvd}(object, ...)
}
\arguments{
\item{object}{Object generated by \code{rrpack}.}

\item{...}{Other argumnts for future usage.}
}
\value{
A numeric matrix.
}
\description{
S3 methods extracting estimated coefficients for objects generated by
\code{rrpack}.
}
