% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrr.R
\name{rrr.cookD}
\alias{rrr.cookD}
\title{Cook's distance in reduced-rank regression for model diagnostics}
\usage{
rrr.cookD(Y, X = NULL, nrank = 1, qr.tol = 1e-07)
}
\arguments{
\item{Y}{response matrix}

\item{X}{covariate matrix}

\item{nrank}{model rank}

\item{qr.tol}{tolerance}
}
\value{
a list containing diagnostics measures
}
\description{
Compute Cook's distance for model diagnostics in \code{rrr} estimation.
}
\references{
Chen, K. Model diagnostics in reduced-rank estimation. \emph{Statistics and
Its interface}, 9, 469--484.
}
