% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{rrr.sim1}
\alias{rrr.sim1}
\title{Simulation model 1}
\usage{
rrr.sim1(n = 50, p = 25, q = 25, nrank = 3, s2n = 1, sigma = NULL,
         rho_X = 0.5, rho_E = 0)
}
\arguments{
\item{n, p, q}{model dimensions}

\item{nrank}{model rank}

\item{s2n}{signal to noise ratio}

\item{sigma}{error variance. If specfied, then s2n has no effect}

\item{rho_X}{correlation parameter in the generation of predictors}

\item{rho_E}{correlation parameter in the generation of random errors}
}
\value{
similated model and data
}
\description{
Similar to the the RSSVD simulation model in Chen, Chan, Stenseth (2012),
JRSSB.
}
\references{
Chen, K., Chan, K.-S. and Stenseth, N. C. (2012) Reduced rank stochastic
regression with a sparse singular value decomposition.  \emph{Journal of the
Royal Statistical Society: Series B}, 74, 203--221.
}
