% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{rrr.sim2}
\alias{rrr.sim2}
\title{Simulation model 2}
\usage{
rrr.sim2(n = 100, p = 50, p0 = 10, q = 50, q0 = 10, nrank = 3, s2n = 1,
         sigma = NULL, rho_X = 0.5, rho_E = 0)
}
\arguments{
\item{n}{sample size}

\item{p}{number of predictors}

\item{p0}{number of relevant predictors}

\item{q}{number of responses}

\item{q0}{number of relevant responses}

\item{nrank}{model rank}

\item{s2n}{signal to noise ratio}

\item{sigma}{error variance. If specfied, then s2n has no effect}

\item{rho_X}{correlation parameter in the generation of predictors}

\item{rho_E}{correlation parameter in the generation of random errors}
}
\value{
similated model and data
}
\description{
Similar to the the SRRR simulation model in Chen and Huang (2012), JASA
}
\references{
Chen, L. and Huang, J.Z. (2012) Sparse reduced-rank regression for
simultaneous dimension reduction and variable selection. \emph{Journal of
the American Statistical Association}. 107:500, 1533--1545.
}
