% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{rrr.sim3}
\alias{rrr.sim3}
\title{Simulation model 3}
\usage{
rrr.sim3(
  n = 100,
  p = 30,
  q.mix = c(5, 20, 5),
  nrank = 2,
  intercept = rep(0.5, 30),
  mis.prop = 0.2
)
}
\arguments{
\item{n}{sample size}

\item{p}{number of predictors}

\item{q.mix}{numbers of Gaussian, Bernolli and Poisson responses}

\item{nrank}{model rank}

\item{intercept}{a vector of intercept}

\item{mis.prop}{missing proportion}
}
\value{
similated model and data
}
\description{
Generate data from a mixed-response reduced-rank regression model
}
\references{
Chen, K., Luo, C., and Liang, J. (2017) Leveraging mixed and incomplete outcomes
through a mixed-response reduced-rank regression. \emph{Technical report}.
}
