% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.mrrr}
\alias{summary.cv.mrrr}
\alias{summary.r4}
\alias{summary.rrr}
\alias{summary.cv.rrr}
\alias{summary.sofar}
\alias{summary.cv.sofar}
\alias{summary.srrr}
\alias{summary.cv.srrr}
\alias{summary.rssvd}
\title{Summarize rrpack Objects}
\usage{
\method{summary}{mrrr}(object, ...)

\method{summary}{cv.mrrr}(object, ...)

\method{summary}{r4}(object, ...)

\method{summary}{rrr}(object, ...)

\method{summary}{cv.rrr}(object, ...)

\method{summary}{sofar}(object, ...)

\method{summary}{cv.sofar}(object, ...)

\method{summary}{srrr}(object, ...)

\method{summary}{cv.srrr}(object, ...)

\method{summary}{rssvd}(object, ...)
}
\arguments{
\item{object}{Object generated from \code{rrpack}.}

\item{...}{Other argumnts for future usage.}
}
\description{
S3 methods summarizing objects generated by \code{rrpack}.
}
